Board is now a new type
not sure whether this is good or not, let's find out.
the idea is to eventually make Board an opaque type
so that we are forced to have all internal impl details
in Core.

but for now it's making the code more verbose, so we still
need to try and see.