scrap Gameplay design.
Scrapping current design because I realized brick expect
clients to provide a single state, not a monad transformer to work with.
We could transform EventM but eventually we'll still need to figure out
how to do the wrapping / unwrapping, might as well switch to a design
that a single data type holds everything for us.